Cleanup (formatting, refactoring), made the project use 4.5.1 which is the lowest possible due to the XMPP lib.
