+++ HarmonyConsole/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using CommandLine;
 using HarmonyHub;
-using HarmonyHub.Entities;
+using HarmonyHub.Entities.Response;
 
 namespace HarmonyConsole
 {
@@ -50,9 +50,9 @@ namespace HarmonyConsole
                 if (!string.IsNullOrEmpty(deviceId) && !string.IsNullOrEmpty(options.Command))
                 {
                     await client.SendKeyPressAsync(deviceId, options.Command);
-				}
+                }
 
-				if (null != harmonyConfig && !string.IsNullOrEmpty(deviceId) && string.IsNullOrEmpty(options.Command))
+                if (null != harmonyConfig && !string.IsNullOrEmpty(deviceId) && string.IsNullOrEmpty(options.Command))
                 {
                     // just list device control options
                     foreach (var device in harmonyConfig.Devices.Where(device => device.Id == deviceId))
@@ -105,9 +105,9 @@ namespace HarmonyConsole
                         }
                     }
                 }
-				Console.WriteLine("Press enter to disconnect");
-				Console.ReadLine();
-			}
-		}
+                Console.WriteLine("Press enter to disconnect");
+                Console.ReadLine();
+            }
+        }
     }
 }
