+++ HarmonyHub/Entities/Response/Config.cs
@@ -2,24 +2,28 @@
 using System.Linq;
 using System.Runtime.Serialization;
 
-namespace HarmonyHub.Entities
+namespace HarmonyHub.Entities.Response
 {
     /// <summary>
-    /// Logitech HarmonyHub Configuration
-    /// </summary>    
+    ///     Logitech HarmonyHub Configuration
+    /// </summary>
     [DataContract]
     public class Config
     {
         [DataMember(Name = "activity")]
         public IList<Activity> Activities { get; set; }
-        [DataMember(Name = "device")]
-        public IList<Device> Devices { get; set; }
-        [DataMember(Name = "sequence")]
-        public IList<object> Sequence { get; set; }
+
         [DataMember(Name = "content")]
         public Content Content { get; set; }
+
+        [DataMember(Name = "device")]
+        public IList<Device> Devices { get; set; }
+
         [DataMember(Name = "global")]
         public Global Global { get; set; }
+
+        [DataMember(Name = "sequence")]
+        public IList<object> Sequence { get; set; }
 
         public string ActivityNameFromId(string activityId)
         {
