+++ HarmonyHub/Entities/Response/ControlGroup.cs
@@ -1,16 +1,16 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-
-namespace HarmonyHub.Entities
+namespace HarmonyHub.Entities.Response
 {
     [DataContract]
     public class ControlGroup
     {
-        [DataMember(Name = "name")]
-        public string Name { get; set; }
         [DataMember(Name = "function")]
         public IList<Function> Functions { get; set; }
+
+        [DataMember(Name = "name")]
+        public string Name { get; set; }
 
         public override string ToString()
         {
