+++ HarmonyHub/Entities/Response/Global.cs
@@ -1,13 +1,14 @@
 using System.Runtime.Serialization;
 
-namespace HarmonyHub.Entities
+namespace HarmonyHub.Entities.Response
 {
     [DataContract]
     public class Global
     {
-        [DataMember(Name = "timeStampHash")]
-        public string TimeStampHash { get; set; }
         [DataMember(Name = "locale")]
         public string Locale { get; set; }
+
+        [DataMember(Name = "timeStampHash")]
+        public string TimeStampHash { get; set; }
     }
 }
